[#5] refactor: type directory import 변경 & totalFeedbackCount로 name 변경에 따른 수정
